+++ json_rpc.go
@@ -8,11 +8,11 @@ import (
 const JSON_RPC_VERSION = "2.0"
 
 // clientRequest represents a JSON-RPC request sent by a client.
-type clientRequest struct {
+type JSONRPCClientRequest struct {
 	// JSON-RPC protocol.
 	Version string `json:"jsonrpc"`
 	// A String containing the name of the method to be invoked.
-	Method string `json:"method"`
+	Method TunManagerRPC `json:"method"`
 	// Object to pass as request parameter to the method.
 	Params interface{} `json:"params,omitempty"`
 	// The request id. This can be of any type. It is used to match the
@@ -21,10 +21,13 @@ type clientRequest struct {
 }
 
 // clientResponse represents a JSON-RPC response returned to a client.
-type ClientResponse struct {
-	Result interface{}   `json:"result"`
-	Error  JSONRPC_Error `json:"error,omitempty"`
-	Id     uint64        `json:"id"`
+type JSONRPCClientResponse struct {
+	// JSON-RPC protocol.
+	Version string `json:"jsonrpc"`
+
+	Result interface{}    `json:"result,omitempty"`
+	Error  *JSONRPC_Error `json:"error,omitempty"`
+	Id     uint64         `json:"id"`
 }
 
 type JSONRPC_ErrorCode int
@@ -49,8 +52,8 @@ const (
 )
 
 // EncodeClientRequest encodes parameters for a JSON-RPC client request.
-func EncodeRPCRequest(method string, args interface{}) ([]byte, error) {
-	c := &clientRequest{
+func EncodeRPCRequest(method TunManagerRPC, args interface{}) ([]byte, error) {
+	c := &JSONRPCClientRequest{
 		Version: JSON_RPC_VERSION,
 		Method:  method,
 		Params:  args,